Move product metadata models to dcpy.models
