+++ dcpy/models/product_metadata.py
