+++ dcpy/test/metadata/test_metadata.py
@@ -5,9 +5,9 @@ import random
 from shapely import wkb, wkt
 import uuid
 
-from dcpy.metadata import models
+import dcpy.models.product_metadata as models
+from dcpy.models.product_metadata import Metadata as md
 from dcpy.metadata import validate
-from dcpy.metadata.models import Metadata as md
 
 METADATA_PATH = (
     Path(__file__).parent.resolve() / "resources" / "test_package" / "metadata.yml"
